production_app/requirements.txt
streamlit>=1.32.0
mlflow>=3.11.1
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0.0
scikit-learn>=1.3.0
PyYAML>=6.0.1
matplotlib>=3.7.0
seaborn>=0.12.0